# Perf: optimize the performance of processMutations