--- a/.changeset/flat-foxes-juggle.md
+++ b/.changeset/flat-foxes-juggle.md
@@ -1,0 +1,5 @@
+---
+'rrweb': patch
+---
+
+Optimize performance of the processMutations