@wfk007 2023-03-23
The dom structure below can be created in many ways:
```
body
    E1
       E2
```

Approach 1:
1. create E1
2. create E2
3. E1 append E2
4. body append E1

Browser Mutation observer callback result:
```
[
    {
        target: body,
        addedNodes: [E1],
        type: "childList",
    },
]
```

approach 2:
1. create E1
2. body append E1
3. create E2
4. E1 append E2

Browser Mutation observer callback result:
```
[
    {
        target: body,
        addedNodes: [E1],
        type: "childList",
    },
    {
        target: E1,
        addedNodes: [E2],
       type: "childList",
    },
]
```

In order to cover different Mutation observer callback results and track adds/removes Nodes, RRweb loops all mutations and its' `addedNodes`.
Each Node in `addedNodes` will still travel to all childNodes.

Considering approach2 result:
1. loop array index 0:
travel E1
travel E2(which is the childNode of E1) 
2. loop array index 1:
travel E2(no need to travel E2 and its' childNodes again)