@Juice10 2023-04-07
Hey @wfk007 this is a great performance improvement, thank you so much for putting all the hard work into this.
One thing I'm missing is making sure your hard work doesn't get reverted by accident in the future.
We have a benchmark suite in the test suite that would be a great candidate for this, especially in combination with a GitHub action (that we don't have yet) https://github.com/benchmark-action/github-action-benchmark that'll make sure performance gets tested on each pull request.