@eoghanmurray 2023-04-14
I haven't looked too closely but the change is nice and small and seems to make sense.
I'm putting this out to a 'canary' environment today to see if any problems arise.
Note: The ESLint failure can be solved by rebasing upon the current trunk.