@wfk007 2023-04-28
@Juice10 @eoghanmurray Hi, I plan to separate this PR into two, thanks for your great suggestions again

part one: https://github.com/rrweb-io/rrweb/pull/1214

part two: is doing